backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sqlalchemy
motor
pymongo
pandas
openpyxl
xlrd==1.2.0
PyJWT
